more casting

+++ zoobot/estimators/bayesian_estimator_funcs.py
@@ -346,8 +346,9 @@ def binomial_loss(labels, predictions):
     ep = 1e-8
     epsilon = tf.constant(ep, dtype=tf.float32)
 
-    yes_votes = labels[0, :]
-    total_votes = labels[1, :]
+    # multiplication in tf requires floats
+    yes_votes = tf.cast(labels[0, :], tf.float32)
+    total_votes = tf.cast(labels[1, :], tf.float32)
     p_yes = tf.identity(predictions)  # fail loudly if passed out-of-range values
 
     # negative log likelihood
